requirements.txt
Django==4.1.7
asgiref==3.6.0
autopep8==2.0.1
dj-database-url==1.2.0
gunicorn==20.1.0
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
toml==0.10.2
whitenoise==6.3.0
django-environ==0.9.0
django-allauth==0.52.0
django-feather==0.3.0
django-sslserver
django-address==0.2.8
django-phonenumber-field
phonenumbers
django-pwa
